@@ -1,9 +1,9 @@
 module github.com/mmcloughlin/avo
 
-go 1.13
+go 1.16
 
 require (
-	golang.org/x/arch v0.0.0-20210923205945-b76863e36670
-	golang.org/x/sys v0.0.0-20211030160813-b3129d9d1021
-	golang.org/x/tools v0.1.7
+	golang.org/x/arch v0.1.0
+	golang.org/x/sys v0.1.0
+	golang.org/x/tools v0.2.0
 )
