@@ -6,10 +6,9 @@ import (
 	"strings"
 
 	"github.com/mmcloughlin/avo/internal/api"
+	"github.com/mmcloughlin/avo/internal/inst"
 	"github.com/mmcloughlin/avo/internal/prnt"
 	"github.com/mmcloughlin/avo/printer"
-
-	"github.com/mmcloughlin/avo/internal/inst"
 )
 
 type ctorstest struct {
