@@ -1,7 +1,6 @@
 // Code generated by command: go run asm.go -out allocfail.s -stubs stubs.go. DO NOT EDIT.
 
 //go:build !appengine && !noasm && gc
-// +build !appengine,!noasm,gc
 
 package allocfail
 
@@ -44,8 +43,9 @@ func encodeBlockAsm12BAvx(dst []byte, src []byte) int
 // emitLiteral writes a literal chunk and returns the number of bytes written.
 //
 // It assumes that:
-//   dst is long enough to hold the encoded bytes
-//   0 <= len(lit) && len(lit) <= math.MaxUint32
+//
+//	dst is long enough to hold the encoded bytes
+//	0 <= len(lit) && len(lit) <= math.MaxUint32
 //
 //go:noescape
 func emitLiteral(dst []byte, lit []byte) int
@@ -53,8 +53,9 @@ func emitLiteral(dst []byte, lit []byte) int
 // emitLiteralAvx writes a literal chunk and returns the number of bytes written.
 //
 // It assumes that:
-//   dst is long enough to hold the encoded bytes
-//   0 <= len(lit) && len(lit) <= math.MaxUint32
+//
+//	dst is long enough to hold the encoded bytes
+//	0 <= len(lit) && len(lit) <= math.MaxUint32
 //
 //go:noescape
 func emitLiteralAvx(dst []byte, lit []byte) int
@@ -68,9 +69,10 @@ func emitRepeat(dst []byte, offset int, length int) int
 // emitCopy writes a copy chunk and returns the number of bytes written.
 //
 // It assumes that:
-//   dst is long enough to hold the encoded bytes
-//   1 <= offset && offset <= math.MaxUint32
-//   4 <= length && length <= 1 << 24
+//
+//	dst is long enough to hold the encoded bytes
+//	1 <= offset && offset <= math.MaxUint32
+//	4 <= length && length <= 1 << 24
 //
 //go:noescape
 func emitCopy(dst []byte, offset int, length int) int
@@ -78,7 +80,8 @@ func emitCopy(dst []byte, offset int, length int) int
 // matchLen returns how many bytes match in a and b
 //
 // It assumes that:
-//   len(a) <= len(b)
+//
+//	len(a) <= len(b)
 //
 //go:noescape
 func matchLen(a []byte, b []byte) int
