@@ -215,20 +215,20 @@ func (s sffxs) Strings() []string {
 }
 
 var sffxsstringsmap = map[sffxs][]string{
-	sffxs{sffxBCST, sffxZ}:   []string{"BCST", "Z"},
-	sffxs{sffxBCST}:          []string{"BCST"},
-	sffxs{sffxRD_SAE, sffxZ}: []string{"RD_SAE", "Z"},
-	sffxs{sffxRD_SAE}:        []string{"RD_SAE"},
-	sffxs{sffxRN_SAE, sffxZ}: []string{"RN_SAE", "Z"},
-	sffxs{sffxRN_SAE}:        []string{"RN_SAE"},
-	sffxs{sffxRU_SAE, sffxZ}: []string{"RU_SAE", "Z"},
-	sffxs{sffxRU_SAE}:        []string{"RU_SAE"},
-	sffxs{sffxRZ_SAE, sffxZ}: []string{"RZ_SAE", "Z"},
-	sffxs{sffxRZ_SAE}:        []string{"RZ_SAE"},
-	sffxs{sffxSAE, sffxZ}:    []string{"SAE", "Z"},
-	sffxs{sffxSAE}:           []string{"SAE"},
-	sffxs{sffxZ}:             []string{"Z"},
-	sffxs{}:                  []string(nil),
+	{sffxBCST, sffxZ}:   {"BCST", "Z"},
+	{sffxBCST}:          {"BCST"},
+	{sffxRD_SAE, sffxZ}: {"RD_SAE", "Z"},
+	{sffxRD_SAE}:        {"RD_SAE"},
+	{sffxRN_SAE, sffxZ}: {"RN_SAE", "Z"},
+	{sffxRN_SAE}:        {"RN_SAE"},
+	{sffxRU_SAE, sffxZ}: {"RU_SAE", "Z"},
+	{sffxRU_SAE}:        {"RU_SAE"},
+	{sffxRZ_SAE, sffxZ}: {"RZ_SAE", "Z"},
+	{sffxRZ_SAE}:        {"RZ_SAE"},
+	{sffxSAE, sffxZ}:    {"SAE", "Z"},
+	{sffxSAE}:           {"SAE"},
+	{sffxZ}:             {"Z"},
+	{}:                  nil,
 }
 
 type sffxscls uint8
@@ -327,54 +327,54 @@ func (i isas) List() []string {
 }
 
 var isaslisttable = [][]string{
-	[]string(nil),
-	[]string{"ADX"},
-	[]string{"SSE2"},
-	[]string{"SSE"},
-	[]string{"SSE3"},
-	[]string{"AES"},
-	[]string{"BMI"},
-	[]string{"SSE4.1"},
-	[]string{"BMI2"},
-	[]string{"CLFLUSH"},
-	[]string{"CLFLUSHOPT"},
-	[]string{"CMOV"},
-	[]string{"CPUID"},
-	[]string{"SSE4.2"},
-	[]string{"AVX512DQ"},
-	[]string{"AVX512BW"},
-	[]string{"AVX512F"},
-	[]string{"LZCNT"},
-	[]string{"MONITOR"},
-	[]string{"MOVBE"},
-	[]string{"SSSE3"},
-	[]string{"PCLMULQDQ"},
-	[]string{"POPCNT"},
-	[]string{"MMX+"},
-	[]string{"RDRAND"},
-	[]string{"RDSEED"},
-	[]string{"RDTSC"},
-	[]string{"RDTSCP"},
-	[]string{"SHA"},
-	[]string{"AVX"},
-	[]string{"AVX512F", "AVX512VL"},
-	[]string{"AES", "AVX"},
-	[]string{"AVX512DQ", "AVX512VL"},
-	[]string{"AVX2"},
-	[]string{"F16C"},
-	[]string{"AVX512VL"},
-	[]string{"AVX512BW", "AVX512VL"},
-	[]string{"AVX512ER"},
-	[]string{"FMA3"},
-	[]string{"AVX512CD", "AVX512VL"},
-	[]string{"AVX512CD"},
-	[]string{"AVX", "PCLMULQDQ"},
-	[]string{"AVX512VBMI", "AVX512VL"},
-	[]string{"AVX512VBMI"},
-	[]string{"AVX512IFMA", "AVX512VL"},
-	[]string{"AVX512IFMA"},
-	[]string{"AVX512VPOPCNTDQ"},
-	[]string{"AVX512BW", "AVX512F"},
+	nil,
+	{"ADX"},
+	{"SSE2"},
+	{"SSE"},
+	{"SSE3"},
+	{"AES"},
+	{"BMI"},
+	{"SSE4.1"},
+	{"BMI2"},
+	{"CLFLUSH"},
+	{"CLFLUSHOPT"},
+	{"CMOV"},
+	{"CPUID"},
+	{"SSE4.2"},
+	{"AVX512DQ"},
+	{"AVX512BW"},
+	{"AVX512F"},
+	{"LZCNT"},
+	{"MONITOR"},
+	{"MOVBE"},
+	{"SSSE3"},
+	{"PCLMULQDQ"},
+	{"POPCNT"},
+	{"MMX+"},
+	{"RDRAND"},
+	{"RDSEED"},
+	{"RDTSC"},
+	{"RDTSCP"},
+	{"SHA"},
+	{"AVX"},
+	{"AVX512F", "AVX512VL"},
+	{"AES", "AVX"},
+	{"AVX512DQ", "AVX512VL"},
+	{"AVX2"},
+	{"F16C"},
+	{"AVX512VL"},
+	{"AVX512BW", "AVX512VL"},
+	{"AVX512ER"},
+	{"FMA3"},
+	{"AVX512CD", "AVX512VL"},
+	{"AVX512CD"},
+	{"AVX", "PCLMULQDQ"},
+	{"AVX512VBMI", "AVX512VL"},
+	{"AVX512VBMI"},
+	{"AVX512IFMA", "AVX512VL"},
+	{"AVX512IFMA"},
+	{"AVX512VPOPCNTDQ"},
+	{"AVX512BW", "AVX512F"},
 }
 
 type opc uint16
